requirements.txt
gspread
google-auth
aiocache
asyncio
python-dotenv
google-cloud-firestore
firebase-admin
openai
aiohttp
beautifulsoup4
tenacity
lxml
lxml_html_clean
pydantic-settings
aioredis
playwright
requests-html
selenium
webdriver-manager
httpx
brotlicffi
python-dateutil
spotipy
pandas
tweepy
twscrape
playwright
snscrape
joblib
scikit-learn
konlpy
nltk
torch
tensorflow
xgboost
lightgbm
rapidfuzz
transformers
sentencepiece
openpyxl
pycountry
statsmodels
pytrends
serpapi
prophet
seaborn
plotly
instaloader
google-search-results